Updated Searcher plugin + translation

+++ src/commands/Searcher/reddit.js
@@ -3,7 +3,7 @@ const { MessageEmbed } = require('discord.js');
 const { KSoftClient } = require('@ksoft/api');
 
 module.exports.run = async (bot, message, args, emojis, settings) => {
-	if (!args[0]) return message.channel.send({ embed:{ color:15158332, description:`${emojis[0]} Please use the format \`${bot.commands.get('reddit').help.usage.replace('${PREFIX}', settings.prefix)}\`.` } }).then(m => m.delete({ timeout: 5000 }));
+	if (!args[0])	return message.error(settings.Language, 'INCORRECT_FORMAT', bot.commands.get('reddit').help.usage.replace('${PREFIX}', settings.prefix)).then(m => m.delete({ timeout: 5000 }));
 	const ksoft = new KSoftClient(bot.config.KSoftSiAPI);
 	try {
 		let reddit;
@@ -19,12 +19,12 @@ module.exports.run = async (bot, message, args, emojis, settings) => {
 			.setTitle(`From /${reddit.post.subreddit}`)
 			.setURL(reddit.post.link)
 			.setImage(reddit.url)
-			.setFooter(`👍 ${reddit.post.upvotes}   👎 ${reddit.post.downvotes} | Provided by KSOFT.API`);
+			.setFooter(`👍 ${reddit.post.upvotes}   👎 ${reddit.post.downvotes} | ${message.translate(settings.Language, 'FUN/MEME_FOOTER')} KSOFT.API`);
 		message.channel.send(embed);
 	} catch (err) {
-		message.delete();
+		if (message.deletable) message.delete();
 		if (bot.config.debug) bot.logger.error(`${err.message} - command: reddit.`);
-		message.channel.send({ embed:{ color:15158332, description:`${emojis[0]} An error occured when running this command, please try again or contact support.` } }).then(m => m.delete({ timeout: 10000 }));
+		message.error(settings.Language, 'ERROR_MESSAGE').then(m => m.delete({ timeout: 5000 }));
 	}
 };
